polishing code structure - game components are now isolated
